+++ Skins/deep/src/components/player/HorizonSlider.vue
@@ -58,7 +58,7 @@ export default {
 </script>
 
 <style scoped lang="scss">
-@import '../index.scss';
+@import '../../index.scss';
 
 .slider-wrap {
   display: inline-block;
